Update modified-run.py
Wait for the overlay to disappear: If the overlaying element is temporary (like a loading spinner), you can wait until the element disappears before clicking the target element.

Scroll to the element: Sometimes, scrolling the element into view can make it clickable. You can do this with element.location_once_scrolled_into_view.

Use JavaScript to click the element: You can use JavaScript to force a click on the element.